requirements.txt
pyspellchecker==0.8.3
streamlit==1.32.0 